# API/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "API",
    "version": "1.0"
  },
  "paths": {
    "/api/Aim/{id}": {
      "get": {
        "tags": [
          "Aim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Aim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAimInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAimInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAimInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Aim"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Aim": {
      "get": {
        "tags": [
          "Aim"
        ],
        "parameters": [
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SourceIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "ClosedOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "SortDescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AimSortBy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Aim"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAimInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAimInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAimInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AimDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Aim/{aimId}/sources/{sourceId}": {
      "post": {
        "tags": [
          "Aim"
        ],
        "parameters": [
          {
            "name": "aimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Aim"
        ],
        "parameters": [
          {
            "name": "aimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Source": {
      "get": {
        "tags": [
          "Source"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceDtoLookup"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceDtoLookup"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceDtoLookup"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Source"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      }
    },
    "/api/Source/{id}": {
      "get": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoDetailed"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoDetailed"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoDetailed"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSourceInput"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSourceInput"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      }
    },
    "/api/Source/archive/{id}": {
      "patch": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      }
    },
    "/api/Source/unarchive/{id}": {
      "patch": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDtoLookup"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AimDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "isClosed": {
            "type": "boolean"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceDtoLookup"
            },
            "nullable": true
          },
          "progress": {
            "$ref": "#/components/schemas/AimProgressDto"
          }
        },
        "additionalProperties": false
      },
      "AimProgressDto": {
        "type": "object",
        "properties": {
          "collectedAmount": {
            "type": "number",
            "format": "double"
          },
          "targetAmount": {
            "type": "number",
            "format": "double"
          },
          "completionPercentage": {
            "type": "number",
            "format": "double",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "AimSortBy": {
        "enum": [
          "Amount",
          "Priority"
        ],
        "type": "string"
      },
...
      "CreateAimInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "currencyId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateSourceInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currencyId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CurrencyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "usdExchangeRate": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "SourceDtoDetailed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "isArchived": {
            "type": "boolean"
          },
          "aims": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AimDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SourceDtoLookup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "isArchived": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateAimInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "priority": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "currencyId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateSourceInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}